#343 refactor: PlaceReviewSummary를 review 패키지로 이동 및 기존 장소 임베딩 초기화 기능 추가
