+++ src/main/java/org/sopt/solply_server/domain/place/service/PlaceSearchDocumentService.java
@@ -6,9 +6,9 @@ import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.sopt.solply_server.domain.place.entity.EmbeddingStatus;
-import org.sopt.solply_server.domain.place.entity.PlaceReviewSummary;
 import org.sopt.solply_server.domain.place.entity.PlaceSearchDocument;
-import org.sopt.solply_server.domain.place.repository.PlaceReviewSummaryRepository;
+import org.sopt.solply_server.domain.review.entity.PlaceReviewSummary;
+import org.sopt.solply_server.domain.review.repository.PlaceReviewSummaryRepository;
 import org.sopt.solply_server.domain.place.repository.PlaceSearchDocumentRepository;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
@@ -21,6 +21,8 @@ public class PlaceSearchDocumentService {
 
     private static final List<EmbeddingStatus> REEMBEDDING_TARGET_STATUSES =
             List.of(EmbeddingStatus.DIRTY, EmbeddingStatus.FAILED);
+    private static final List<EmbeddingStatus> INIT_TARGET_STATUSES =
+            List.of(EmbeddingStatus.INIT);
 
     private final PlaceSearchDocumentRepository placeSearchDocumentRepository;
     private final PlaceReviewSummaryRepository placeReviewSummaryRepository;
@@ -42,6 +44,19 @@ public class PlaceSearchDocumentService {
             document.markDirty();
             log.info("PlaceSearchDocument DIRTY 전환 (태그 변경) - placeId={}, tagId={}", document.getPlaceId(), tagId);
         });
+    }
+
+    public void initializePendingDocuments() {
+        List<PlaceSearchDocument> pendingDocuments = placeSearchDocumentRepository
+                .findAllByStatusIn(INIT_TARGET_STATUSES);
+
+        log.info("신규 임베딩 초기화 시작 - 대상 수: {}", pendingDocuments.size());
+
+        for (PlaceSearchDocument document : pendingDocuments) {
+            batchProcessor.processOne(document.getPlaceId(), null);
+        }
+
+        log.info("신규 임베딩 초기화 완료");
     }
 
     @Scheduled(cron = "0 0 3 * * *") // 매일 새벽 3시
