+++ src/main/java/org/sopt/solply_server/domain/review/repository/PlaceReviewSummaryRepository.java
@@ -1,6 +1,6 @@
-package org.sopt.solply_server.domain.place.repository;
+package org.sopt.solply_server.domain.review.repository;
 
-import org.sopt.solply_server.domain.place.entity.PlaceReviewSummary;
+import org.sopt.solply_server.domain.review.entity.PlaceReviewSummary;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PlaceReviewSummaryRepository extends JpaRepository<PlaceReviewSummary, Long> {
